requirements.txt
asgiref==3.9.1
audioop-lts==0.2.1
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.3
cycler==0.12.1
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.9.0
django-jazzmin==3.0.1
django-semantic-forms==0.1.8
djangorestframework==3.16.1
docopt==0.6.2
filelock==3.19.1
fonttools==4.60.1
graphviz==0.20.3
gunicorn==23.0.0
idna==3.10
Js2Py==0.74
kiwisolver==1.4.9
matplotlib==3.10.7
mysqlclient==2.2.7
numpy==2.2.5
packaging==25.0
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.3.8
psycopg2-binary==2.9.11
PyAudio==0.2.14
pyjsparser==2.7.1
PyMySQL==1.1.2
pyparsing==3.2.5
PyPrind==2.11.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
PythonTurtle==0.3.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
SpeechRecognition==3.14.3
sqlparse==0.5.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
virtualenv==20.34.0
whitenoise==6.11.0
